+++ src/Attributes/Unauthenticated.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Knuckles\Scribe\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_FUNCTION | Attribute::TARGET_METHOD | Attribute::TARGET_CLASS)]
+class Unauthenticated
+{
+    public function toArray()
+    {
+        return ["authenticated" => false];
+    }
+}