+++ src/ScribeServiceProvider.php
@@ -13,48 +13,15 @@ use Knuckles\Scribe\Tools\Utils;
 
 class ScribeServiceProvider extends ServiceProvider
 {
-    /**
-     * Bootstrap the application events.
-     *
-     * @return void
-     */
     public function boot()
     {
-        // Register custom Markdown Blade compiler so we can automatically have MD views converted to HTML
-        $this->app->view->getEngineResolver()
-            ->register('blademd', fn() => new BladeMarkdownEngine($this->app['blade.compiler']));
-        $this->app->view->addExtension('md.blade.php', 'blademd');
+        $this->registerViews();
 
-        $this->loadViewsFrom(__DIR__ . '/../resources/views/', 'scribe');
-
-        // Publish views in separate, smaller groups for ease of end-user modifications
-        $viewGroups = [
-            'views' => '',
-            'examples' => 'partials/example-requests',
-            'themes' => 'themes',
-            'markdown' => 'markdown',
-        ];
-        foreach ($viewGroups as $group => $path) {
-            $this->publishes([
-                __DIR__ . "/../resources/views/$path" => $this->app->basePath("resources/views/vendor/scribe/$path"),
-            ], "scribe-$group");
-        }
-
-        $this->publishes([
-            __DIR__ . '/../config/scribe.php' => $this->app->configPath('scribe.php'),
-        ], 'scribe-config');
-
-        $this->mergeConfigFrom(__DIR__ . '/../config/scribe.php', 'scribe');
+        $this->registerConfig();
 
         $this->bootRoutes();
 
-        if ($this->app->runningInConsole()) {
-            $this->commands([
-                GenerateDocumentation::class,
-                MakeStrategy::class,
-                Upgrade::class,
-            ]);
-        }
+        $this->registerCommands();
 
         // Bind the route matcher implementation
         $this->app->bind(RouteMatcherInterface::class, config('scribe.routeMatcher', RouteMatcher::class));
@@ -76,6 +43,49 @@ class ScribeServiceProvider extends ServiceProvider
         ) {
             $routesPath = Utils::isLumen() ? __DIR__ . '/../routes/lumen.php' : __DIR__ . '/../routes/laravel.php';
             $this->loadRoutesFrom($routesPath);
+        }
+    }
+
+    protected function registerViews(): void
+    {
+        // Register custom Markdown Blade compiler so we can automatically have MD views converted to HTML
+        $this->app->view->getEngineResolver()
+            ->register('blademd', fn() => new BladeMarkdownEngine($this->app['blade.compiler']));
+        $this->app->view->addExtension('md.blade.php', 'blademd');
+
+        $this->loadViewsFrom(__DIR__ . '/../resources/views/', 'scribe');
+
+        // Publish views in separate, smaller groups for ease of end-user modifications
+        $viewGroups = [
+            'views' => '',
+            'examples' => 'partials/example-requests',
+            'themes' => 'themes',
+            'markdown' => 'markdown',
+        ];
+        foreach ($viewGroups as $group => $path) {
+            $this->publishes([
+                __DIR__ . "/../resources/views/$path" => $this->app->basePath("resources/views/vendor/scribe/$path"),
+            ], "scribe-$group");
+        }
+    }
+
+    protected function registerConfig(): void
+    {
+        $this->publishes([
+            __DIR__ . '/../config/scribe.php' => $this->app->configPath('scribe.php'),
+        ], 'scribe-config');
+
+        $this->mergeConfigFrom(__DIR__ . '/../config/scribe.php', 'scribe');
+    }
+
+    protected function registerCommands(): void
+    {
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                GenerateDocumentation::class,
+                MakeStrategy::class,
+                Upgrade::class,
+            ]);
         }
     }
 }